changed package mep to keyword
Former-commit-id: 827bed5a8cc0ca30b21435cb06b05cc5f8b765d8

+++ server/admission/de.uniol.inf.is.odysseus.admission.status/src/de/uniol/inf/is/odysseus/admission/status/mep/ChooseLoadSheddingGrowthFunction.java
@@ -15,7 +15,7 @@ public class ChooseLoadSheddingGrowthFunction extends AbstractFunction<Command> 
 
 	@Override 
 	public Command getValue() {
-		final int growth = (int) getInputValue(0);
+		final int growth = ((Long) getInputValue(0)).intValue();
 		
 		return new Command() {
 			@Override public boolean run() {
